src/backend/requirements.txt
beautifulsoup4==4.6.0
Django==1.11.7
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.3.0
django-filter==1.1.0
djangorestframework==3.8.2
drf-nested-routers==0.90.0
drf_yasg==1.8.0
elasticsearch==6.2.0
flex==6.13.2
future==0.16.0
lxml==4.2.1
pandas==0.23.0
requests==2.18.4
swagger-spec-validator==2.3.1
